Replace win32-service dependency
GCC builds of Ruby are not compatible with win32-service due
the usage of SEH inside deamons.

To workaround that, instead of using win32-service gems, I've
replaced the code with a custom module that mimics the functionality
and stays pure-ruby. It shells out to 'sc' utility, which is part
of every Windows installation.